Part 25- Adding Identity To Project

+++ Core/Adly.Application/Contracts/User/Models/JwtAccessTokenModel.cs
@@ -1,3 +1,3 @@
 ﻿namespace Adly.Application.Contracts.User.Models;
 
-public record JwtAccessTokenModel(string AccessToken,int ExpireInSeconds, string TokenType="Bearer");
+public record JwtAccessTokenModel(string AccessToken,double ExpireInSeconds, string TokenType="Bearer");
